requirements.txt
streamlit
ollama
chromadb
PyMuPDF
langchain-text-splitters
groq
python-dotenv
